+++ jinns/nn/_hyperpinn.py
@@ -11,7 +11,7 @@ from typing import Callable, Literal, Self, Union, Any, cast
 from math import prod
 import jax
 import jax.numpy as jnp
-from jaxtyping import Array, Float, PyTree, Key
+from jaxtyping import PRNGKeyArray, Array, Float, PyTree
 import equinox as eqx
 import numpy as onp
 
@@ -213,12 +213,13 @@ class HyperPINN(PINN):
     @classmethod
     def create(
         cls,
+        *,
         eq_type: Literal["ODE", "statio_PDE", "nonstatio_PDE"],
         hyperparams: list[str],
         hypernet_input_size: int,
+        key: PRNGKeyArray | None = None,
         eqx_network: eqx.nn.MLP | MLP | None = None,
         eqx_hyper_network: eqx.nn.MLP | MLP | None = None,
-        key: Key = None,
         eqx_list: tuple[tuple[Callable, int, int] | tuple[Callable], ...] | None = None,
         eqx_list_hyper: (
             tuple[tuple[Callable, int, int] | tuple[Callable], ...] | None
@@ -370,10 +371,10 @@ class HyperPINN(PINN):
 
             ### Now we finetune the hypernetwork architecture
 
-            key, subkey = jax.random.split(key, 2)
+            subkey1, subkey2 = jax.random.split(key, 2)
             # with warnings.catch_warnings():
             #    warnings.filterwarnings("ignore", message="A JAX array is being set as static!")
-            eqx_network = MLP(key=subkey, eqx_list=eqx_list)
+            eqx_network = MLP(key=subkey1, eqx_list=eqx_list)
             # quick partitioning to get the params to get the correct number of neurons
             # for the last layer of hyper network
             params_mlp, _ = eqx.partition(eqx_network, eqx.is_inexact_array)
@@ -416,10 +417,9 @@ class HyperPINN(PINN):
                         + eqx_list_hyper[2:]
                     ),
                 )
-            key, subkey = jax.random.split(key, 2)
             # with warnings.catch_warnings():
             #    warnings.filterwarnings("ignore", message="A JAX array is being set as static!")
-            eqx_hyper_network = cast(MLP, MLP(key=subkey, eqx_list=eqx_list_hyper))
+            eqx_hyper_network = cast(MLP, MLP(key=subkey2, eqx_list=eqx_list_hyper))
 
             ### End of finetuning the hypernetwork architecture