+++ jinns/nn/_ppinn.py
@@ -10,7 +10,7 @@ import jax
 import jax.numpy as jnp
 import equinox as eqx
 
-from jaxtyping import Array, Key, Float
+from jaxtyping import Array, Float, PRNGKeyArray
 
 from jinns.parameters._params import Params
 from jinns.nn._pinn import PINN
@@ -123,9 +123,10 @@ class PPINN_MLP(PINN):
     @classmethod
     def create(
         cls,
+        *,
+        key: PRNGKeyArray | None = None,
         eq_type: Literal["ODE", "statio_PDE", "nonstatio_PDE"],
         eqx_network_list: list[eqx.nn.MLP | MLP] | None = None,
-        key: Key = None,
         eqx_list_list: (
             list[tuple[tuple[Callable, int, int] | tuple[Callable], ...]] | None
         ) = None,
@@ -213,7 +214,7 @@ class PPINN_MLP(PINN):
 
             eqx_network_list = []
             for eqx_list in eqx_list_list:
-                key, subkey = jax.random.split(key, 2)
+                key, subkey = jax.random.split(key, 2)  # type: ignore
                 eqx_network_list.append(MLP(key=subkey, eqx_list=eqx_list))
 
         ppinn = cls(