+++ docs/changelog.md
@@ -5,6 +5,7 @@
 * v1.6.0
 
     - `Params.eq_params` from `dict` to `eqx.Module` **internally** [!85](https://gitlab.com/mia_jinns/jinns/-/merge_requests/85). A list of (limited) breaking changes for the user is given at the MR link. Related issue [#22](https://gitlab.com/mia_jinns/jinns/-/issues/22)
+    - Improve type hints for dataclasses by writing `__init__` methods when needed [!87](https://gitlab.com/mia_jinns/jinns/-/merge_requests/87)
     - Fix wrong imports in load/save demo notebooks. [See this issue on Github](https://github.com/mia-jinns/jinns/issues/1)